+++ internal/s3/__init__.py
@@ -1,0 +1,11 @@
+import click
+
+from internal.s3.ls import ls
+
+
+@click.group(context_settings={"help_option_names": ["-h", "--help"]})
+def cli():
+    """S3 utilities."""
+
+
+cli.add_command(ls)